per-directory external database list
indexer (probably admin) no longer has to keep track of external database basenames that users want to track
directory owner can create a file called external.gufi listing all external db files to track
changed -q to trigger validating external db files before tracking them
    defaults to not set

+++ include/utils.h
@@ -106,9 +106,9 @@ int shortpath(const char *name, char *nameout, char *endname);
 /* descend */
 typedef int (*process_nondir_f)(struct work *nondir, struct entry_data *ed, void *nondir_args);
 
-/* returns 1 for success, 0 for failure */
-typedef int (*process_external_db_f)(struct input *in,
-                                     struct work *child, void *args);
+/* returns number of external databases tracked */
+typedef size_t (*process_external_db_f)(struct input *in,
+                                        struct work *child, void *args);
 
 struct descend_counters {
     size_t dirs;
